requirements.txt
Django==4.0.4
django-cors-headers==3.13.0
django-shortcuts==1.6
djangorestframework==3.13.1
drf-yasg==1.20.0
flake8==4.0.1
python-decouple==3.6
pika==1.3.1